feat(sse): add SSE event streamer endpoint
Add new /sse endpoint for streaming system events (Server-Sent Events).
Events published: OperatorStarted, OperatorStopped, AgentConnected,
AgentDisconnected, CycleStarted, CycleFinished.

- Add events.go with SSE event types and payloads
- Add /sse handler to api_server.go (SSE format streaming)
- Add event ring buffer to orchestrator for SSE events
- Add publishEvent helper to publish events to SSE clients
- Add SSE metrics (SSEClientsConnected, SSEConnectionsTotal, etc.)
- Update scheduler to publish cycle events

+++ internal/orchestrator/events.go
@@ -46,13 +46,13 @@ type AgentDisconnectedData struct {
 
 // CycleStartedData is the payload for CycleStarted events.
 type CycleStartedData struct {
-	Cycle uint64 `json:"cycle"`
+	Cycle int `json:"cycle"`
 }
 
 // CycleFinishedData is the payload for CycleFinished events.
 type CycleFinishedData struct {
-	Cycle  uint64 `json:"cycle"`
-	Issued uint64 `json:"issued"`
+	Cycle  int `json:"cycle"`
+	Issued int `json:"issued"`
 }
 
 // MarshalJSON implements json.Marshaler for SSEEvent.
